+++ src/auth/auth.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
-import { ONEDAY } from './jwt/jwt.payload';
-import { JwtStrategy } from './jwt/jwt.strategy';
+import { ONEDAY } from '../common/decorator/auth/jwt/jwt.payload';
+import { JwtStrategy } from '../common/decorator/auth/jwt/jwt.strategy';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { User } from '../users/entities/users.entity';
